--- src/main/java/study/racingcar/controller/GameController.java
@@ -1,27 +1,0 @@
-package study.racingcar.controller;
-
-import study.racingcar.view.InputView;
-import study.racingcar.view.OutputView;
-import study.racingcar.domain.RacingCarGame;
-
-public class GameController {
-    /**
-     * 게임 실행을 담당한다
-     */
-    public static void main(String[] args) {
-        final int numOfCar = getRoundsToPlay();
-        final int rounds = getTotalNumberOfCar();
-        RacingCarGame racingCarGame = new RacingCarGame(numOfCar, rounds);
-        racingCarGame.run();
-    }
-
-    private static int getRoundsToPlay() {
-        OutputView.printTryCountSign();
-        return InputView.getTryCount();
-    }
-
-    private static int getTotalNumberOfCar() {
-        OutputView.printCarNumberSign();
-        return InputView.getCarNumber();
-    }
-}
